requirements.txt
numpy
biopython
hmmlearn
scikit-learn
matplotlib
joblib
seaborn

